Revert "Feat/testing (#5)"
This reverts commit b33230e9038aee3af6f3a79da013504fecd56386.
