+++ go.mod
@@ -1,32 +1,30 @@
-module github.com/vshn/appcat-cli
+module github.com/vshn/appcat-cmd
 
 go 1.20
 
-require k8s.io/apimachinery v0.26.3 // indirect
+require k8s.io/apimachinery v0.26.2 // indirect
+
+require k8s.io/api v0.26.2 // indirect
 
 require (
-	github.com/stretchr/testify v1.8.2 // indirect
-	k8s.io/api v0.26.3 // indirect
-)
-
-require (
-	github.com/go-logr/logr v1.2.4 // indirect
+	github.com/fatih/structs v1.1.0
+	github.com/go-logr/logr v1.2.3 // indirect
 	github.com/gogo/protobuf v1.3.2 // indirect
 	github.com/google/gofuzz v1.2.0 // indirect
 	github.com/json-iterator/go v1.1.12 // indirect
 	github.com/modern-go/concurrent v0.0.0-20180306012644-bacd9c7ef1dd // indirect
 	github.com/modern-go/reflect2 v1.0.2 // indirect
 	github.com/sirupsen/logrus v1.9.0
-	github.com/vshn/component-appcat/apis v0.0.0-20230411090651-b0418f5ceb5c
-	golang.org/x/net v0.8.0 // indirect
-	golang.org/x/sys v0.7.0 // indirect
-	golang.org/x/text v0.8.0 // indirect
+	github.com/vshn/component-appcat/apis v0.0.0-20230310100927-7153c56890ce
+	golang.org/x/net v0.7.0 // indirect
+	golang.org/x/sys v0.5.0 // indirect
+	golang.org/x/text v0.7.0 // indirect
 	gopkg.in/inf.v0 v0.9.1 // indirect
 	gopkg.in/yaml.v2 v2.4.0 // indirect
-	k8s.io/klog/v2 v2.90.1 // indirect
-	k8s.io/utils v0.0.0-20230313181309-38a27ef9d749 // indirect
+	k8s.io/klog/v2 v2.80.1 // indirect
+	k8s.io/utils v0.0.0-20221128185143-99ec85e7a448 // indirect
 	sigs.k8s.io/controller-runtime v0.14.1 // indirect
-	sigs.k8s.io/json v0.0.0-20221116044647-bc3834ca7abd // indirect
+	sigs.k8s.io/json v0.0.0-20220713155537-f223a00ba0e2 // indirect
 	sigs.k8s.io/structured-merge-diff/v4 v4.2.3 // indirect
 	sigs.k8s.io/yaml v1.3.0
 )
