--- tests/golden/exoKafka/default.yaml
@@ -1,20 +1,0 @@
-apiVersion: exoscale.appcat.vshn.io/v1
-kind: ExoscaleKafka
-metadata:
-  creationTimestamp: null
-  generateName: 'my-kafka-example '
-  namespace: my-namespace
-spec:
-  parameters:
-    maintenance:
-      dayOfWeek: Sunday
-      timeOfDay: "24:00:00"
-    network: {}
-    service:
-      kafkaSettings: null
-      version: 3.4.0
-      zone: ch-dk-2
-    size:
-      plan: 'startup-2 '
-  writeConnectionSecretToRef: {}
-status: {}
