--- tests/golden/exoKafka/parameters.yaml
@@ -1,20 +1,0 @@
-apiVersion: exoscale.appcat.vshn.io/v1
-kind: ExoscaleKafka
-metadata:
-  creationTimestamp: null
-  generateName: 'my-kafka-example '
-  namespace: my-namespace
-spec:
-  parameters:
-    maintenance:
-      dayOfWeek: Sunday
-      timeOfDay: "24:00:00"
-    network: {}
-    service:
-      kafkaSettings: null
-      version: 3.4.0
-      zone: ch-fr-1
-    size:
-      plan: enterprise-large
-  writeConnectionSecretToRef: {}
-status: {}
