--- tests/golden/exoMySQL/parameters.yaml
@@ -1,22 +1,0 @@
-apiVersion: exoscale.appcat.vshn.io/v1
-kind: ExoscaleMySQL
-metadata:
-  creationTimestamp: null
-  generateName: my-mysql-example
-  namespace: my-namespace
-spec:
-  parameters:
-    backup:
-      timeOfDay: "24:00:00"
-    maintenance:
-      dayOfWeek: Sunday
-      timeOfDay: "24:00:00"
-    network: {}
-    service:
-      majorVersion: "9999"
-      mysqlSettings: null
-      zone: ch-dk-2
-    size:
-      plan: 'hobbyist-2 '
-  writeConnectionSecretToRef: {}
-status: {}
