--- tests/golden/exoOpenSearch.json
@@ -1,4 +1,0 @@
-{
-    "default": "ExoscaleOpenSearch",
-    "parameters": "ExoscaleOpenSearch --spec.parameters.backup.timeOfDay=24:00:00 spec.parameters.size.plan enterprise-large --spec.parameters.service.majorVersion 9999"
-}
