--- tests/golden/exoOpenSearch/default.yaml
@@ -1,22 +1,0 @@
-apiVersion: exoscale.appcat.vshn.io/v1
-kind: ExoscaleOpenSearch
-metadata:
-  creationTimestamp: null
-  generateName: my-openSearch-example
-  namespace: my-namespace
-spec:
-  parameters:
-    backup:
-      timeOfDay: "12:00:00"
-    maintenance:
-      dayOfWeek: Sunday
-      timeOfDay: "24:00:00"
-    network: {}
-    service:
-      majorVersion: "2"
-      opensearchSettings: null
-      zone: ch-dk-2
-    size:
-      plan: 'hobbyist-2 '
-  writeConnectionSecretToRef: {}
-status: {}
