--- tests/golden/exoPostgreSQL/default.yaml
@@ -1,22 +1,0 @@
-apiVersion: exoscale.appcat.vshn.io/v1
-kind: ExoscalePostgreSQL
-metadata:
-  creationTimestamp: null
-  generateName: my-postgres-example
-  namespace: my-namespace
-spec:
-  parameters:
-    backup:
-      timeOfDay: "12:00:00"
-    maintenance:
-      dayOfWeek: Sunday
-      timeOfDay: "24:00:00"
-    network: {}
-    service:
-      majorVersion: "15"
-      pgSettings: null
-      zone: ch-dk-2
-    size:
-      plan: 'hobbyist-2 '
-  writeConnectionSecretToRef: {}
-status: {}
