--- tests/golden/exoPostgreSQL/parameters.yaml
@@ -1,22 +1,0 @@
-apiVersion: exoscale.appcat.vshn.io/v1
-kind: ExoscalePostgreSQL
-metadata:
-  creationTimestamp: null
-  generateName: my-postgres-example
-  namespace: my-namespace
-spec:
-  parameters:
-    backup:
-      timeOfDay: "24:00:00"
-    maintenance:
-      dayOfWeek: Sunday
-      timeOfDay: "24:00:00"
-    network: {}
-    service:
-      majorVersion: "13"
-      pgSettings: null
-      zone: ch-dk-2
-    size:
-      plan: enterprise-large
-  writeConnectionSecretToRef: {}
-status: {}
