--- tests/golden/exoRedis/default.yaml
@@ -1,18 +1,0 @@
-apiVersion: exoscale.appcat.vshn.io/v1
-kind: ExoscaleRedis
-metadata:
-  creationTimestamp: null
-  generateName: redis-app1-prod
-  namespace: prod-app
-spec:
-  parameters:
-    maintenance:
-      dayOfWeek: Sunday
-      timeOfDay: "24:00:00"
-    network: {}
-    service:
-      redisSettings: null
-      zone: ch-dk-2
-    size: {}
-  writeConnectionSecretToRef: {}
-status: {}
