--- tests/golden/vshnPostgreSQL.json
@@ -1,5 +1,0 @@
-{
-    "default": "VSHNPostgreSQL",
-    "parameters": "VSHNPostgreSQL --spec.parameters.backup.retention=9999 --spec.parameters.backup.schedule \"1 1 1 1 1\" --spec.parameters.size.CPU 100000m"
-}
-
