--- tests/golden/vshnPostgreSQL/default.yaml
@@ -1,29 +1,0 @@
-apiVersion: vshn.appcat.vshn.io/v1
-kind: VSHNPostgreSQL
-metadata:
-  creationTimestamp: null
-  generateName: pgsql-app1-prod
-  namespace: prod-app
-spec:
-  parameters:
-    backup:
-      retention: 12
-      schedule: 30 23 * * *
-    maintenance: {}
-    network: {}
-    restore: {}
-    scheduling:
-      nodeSelector:
-        appuio.io/node-class: plus
-    service:
-      majorVersion: "15"
-      pgSettings: null
-    size:
-      cpu: 600m
-      disk: 80Gi
-      memory: 3500Mi
-      requests:
-        cpu: 300m
-        memory: 1000Mi
-  writeConnectionSecretToRef: {}
-status: {}
