--- tests/golden/vshnRedis/parameters.yaml
@@ -1,22 +1,0 @@
-apiVersion: vshn.appcat.vshn.io/v1
-kind: VSHNRedis
-metadata:
-  creationTimestamp: null
-  generateName: redis-app1-prod
-  namespace: prod-app
-spec:
-  parameters:
-    scheduling: {}
-    service:
-      redisSettings: '|activedefrag yes'
-      version: "9999"
-    size:
-      cpuLimits: 1000m
-      cpuRequests: 500m
-      disk: 80Gi
-      memoryLimits: 1Gi
-      memoryRequests: 500Mi
-    tls: {}
-  writeConnectionSecretToRef:
-    name: redisSecret
-status: {}
